+++ packages/cli/plugin-analyze/tests/fixtures/entries/index-entry/src/index.jsx
@@ -1,8 +1,0 @@
-/* eslint-disable no-undef */
-import React from 'react';
-import { Render } from 'react-dom';
-
-const App = () => <div>app</div>;
-
-Render(<App />, document.getElementById('root'));
-/* eslint-enable no-undef */