+++ packages/cli/plugin-bff/modern.config.js
@@ -1,6 +1,2 @@
 /** @type {import('@modern/module-tools').UserConfig} */
-module.exports = {
-  output: {
-    packageMode: 'node-js',
-  },
-};
+module.exports = {};