+++ packages/solutions/app-tools/tests/analyze/fixtures/entries/index-entry/src/index.jsx
@@ -1,0 +1,10 @@
+/* eslint-disable no-undef */
+// eslint-disable-next-line no-unused-vars
+import React from 'react';
+import { Render } from 'react-dom';
+
+// eslint-disable-next-line no-unused-vars
+const App = () => <div>app</div>;
+
+Render(<App />, document.getElementById('root'));
+/* eslint-enable no-undef */