+++ packages/cli/plugin-ssg/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/plugin-ssg
 
+## 1.2.16
+
+### Patch Changes
+
+- 5876e63: fix: the SSG callback is called once for each baseUrl, when a user configures multiple baseUrl.
+  fix: 当用户配置了多个 baseUrl 时，将会为每个 baseUrl 调用一次 SSG 回调函数。
+- Updated dependencies [4fc801f]
+- Updated dependencies [c8614b8]
+  - @modern-js/utils@1.8.0
+
 ## 1.2.15
 
 ### Patch Changes